Code for adding data to sqlite database
This is supposed to avoid the bugs that were coming up with SoundCloud
objects missing particular attributes: it adds 'NULL' if an attribute
is missing. But it's also supposed to reduce headaches in another way,
because it generalises the code for adding rows to tables in the
database. You need to create two strings per table, one used for the
SQL statement that creates the table, and one which contains all the
attributes that will be put into columns in the table (this will be the
same as the previous string, minus all the capital letters! I guess I
could have written a regex to convert one to the other, but… it's late).