Removed need for two strings per table in add_data.py
I couldn't stand thinking about all that redundancy. So now there's a
simple function called att_string() that turns the string containing
the bit of SQL that defines the columns in the table into a string that
defines which attributes we want to get from the SoundCloud data
objects and put into columns in the SQLite table. Fewer potential
mistakes!