Expanded into potential replacement for chunks of getSoundCloudData.py
Full comments added, plus a (hopefully) faster version of the function
for adding data to tables. Strings for creating (and populating) the
tables we'll need have now been added, based on what's in
getSoundCloudData.py. I hope that some big chunks of
getSoundCloudData.py will be possible to remove now, replacing them
with calls to functions (and the aforementioned strings) here. Having
all this stuff in a separate module may help to avoid bugs.