New genres & tags script, add_data.py tidied
Added a script that will generate the analysis of genres & tags
presented in the genres2.doc file now in the data folder of our
Dropbox. Also tidied up add_data.py so that the table-creator strings
are packaged up in a dict and create_table and insert_data functions
can now be called with cursor, data, and table names only (the latter
of which they will look up in that dict).