genre_relationships.py in fully working order
This now stores all its output in an sqlite database (usable for
further network analysis etc). Instead of requiring the NLTK, now
includes the stopwords from it as a pickled list. Changes to
add_data.py so that it can now do all the database writing to both
sqlite databases.