comments_corp.py now in fully working order
comments_corp.py now produces a table in the _deriv database that can
be queried to assemble corpora for comparative analysis. Corpus text
(lightly preprocessed) is in the last column; first column gives the id
from which the original data can be found; other columns contain
metadata variables to refer to in SQL queries when automatically
assembling corpora from the text. See comments at start of file for
explanation.