Minimal snowball sampler now works
This now collects data and saves it to an SQLite database. It uses a
different approach than get_soundcloud_data: first, it only requests
followers and followings, and bases the users and x_follows_y on those,
ignoring anything else; second, it works in 'steps', basically the
number of degrees of separation from a starting individual. The idea is
that because we only get up to 50 follows and followings, we may need
to keep re-sampling from the same starting points, but without
necessarily wanting to go many degrees of separation away from them
until we've covered their immediate networks (say, when the number of
followers for each individual is higher than a certain proportion of
the total that exist). So the next thing will be to write the code that
queries the database and decides who to re-request data on. Also to
start calling some of the other helpful functions in
get_soundcloud_data.py, for example to introduce a time delay. At the
moment, I've tried running this with just two steps from one of our
interviewees, just once - I got 4264 follow relationships between 3185
individuals with 101 calls to the Soundcloud API, which took well under
a minute (the max would have been 10001 individuals, in the same amount
of time).