forward_waves.py now seems to work!
This is a new program for crawling SoundCloud. It only crawls forward
from follower to followed, and it does so in waves, each corresponding
to a specific degree of separation from a seed individual. It collects
all individuals in a single wave before moving to the next. Holds as
little data as possible in memory and as much as possible in .tmp files
and the sqlite database to which it's adding. Data collected is held in
buffers of a certain size; when they're filled, they're cleared and
contents is sent to the sqlite database. If interrupted, has a resume
function for picking up where it left off.