Collecting data to show mutual commenting
comment_interactions.py finds the most frequent commenters on a given
user's tracks, plus the users that user most frequently comments on,
and then collects all comments on all tracks by all these people. The
point of this is to get comment data that's likely to include lots of
conversations: both for visualisation and for linguistic analysis.