rand_samp.py now backs up the databases
rand_samp.py is now a bit more like get_soundcloud_data.py in that it
backs up its database and pauses between batches. Batches are of 1000,
and take about half an hour to collect. It would be speeded up by
modifications to all_data_one_person.py to prevent re-collections of
data on individual tracks and individual users but I haven't had time
to do that yet. Maybe best to run it as it is, and leave it running.