Adjustments to the analysis of genre
genre_network2.py is basically re-written and now produces output
targeted for Gephi (accommodating some Gephi weirdness);
genre_relationships_sample.py is like genre_relationships.py except
that it takes account of the fact that there is a conceptual difference
between users in the 'sample' and everyone else in the database. At the
moment, it takes account of that by analysing only tracks created by
users in the sample; I'll be adjusting that soon so that it analyses
all of them, but only includes those created by users in the sample in
some metrics (esp. genre popularity).